Adds initial cut of CPU
This is an initial cut of the CPU, working with verilator. Make sure
you have verilator and cmake installed on your platform. On Fedora, I
had to also install the verilator-devel package so cmake would work.

The CPU is functional, though currently fills the pipeline with no-ops
instead of implementing any hazard management.