Added register hazard management
All writes to general registers are now taken into account in the
pipeline, so the result is passed backwards in the pipeline to previous
instructions if those instructions rely on that result.

There's three kinds of hazards solved within this commit, and all three
are mainly to do with instructions needing to use results that are still
being saved into registers: ALU hazards, memory load hazards, store
hazard.

ALU hazards are when the ALU result is used in the next ALU instruction.
The result of the ALU is simply used instead of one of the registers in
the next instruction.

A memory load hazard is when a result from a load is used in the next
ALU instruction. The result from the load is fed directly back into the
ALU.

A store hazard is when a store occurs with a register whose value hasn't
been saved yet. That value is fed directly to the mem_access module for
it to save to memory, before the value has actually hit the register
bank.