Added branching and branch hazard management
This commit adds support for branching as well as the associated hazard
management.

There's two main issues to look out for:
1) Jumping to a register that was just modified
2) Prediction and solving mis-predictions

The easy one is jumping to a register that was just modified. We detect
during decoding if a branch is happening to a register that is being
modified in either of the two previous instructions. If so, the result
of that particular stage of the pipeline is used instead of the register
itself. This is untested, as I have no code yet which performs a jump to
register.

For prediction, we adopt an "always branch" approach to prediction. Most
of the time, it's how loops work which is where we are likely to see
gains the most. Then, the ALU must send a flag backwards to the fetcher
to inform it to revert its PC back to what it was if the branch should
not have been taken, and the flag is also sent to the decoder to mark
the instruction as invalid before it continues down the pipeline, but
essentially as a no-op.

Another major caveat is that while much of the work for jump and link
from an instruction fetching and execution point of view is there, the
"and link" bit is not implemented. That will come later.

The classic test program executes successfully, having executed 46
actual instructions in 57 clocks.